+++ environment.yml
@@ -4,7 +4,7 @@ channels:
 dependencies:
   - python=3.12
   - numpy
+  - row
   - signac
-  - signac-flow
   - signac-dashboard
